auth started

+++ src/controller/dealer.controller.ts
@@ -5,12 +5,16 @@ import DealerModel, { DealerDocument } from "../models/dealer.model";
 
 const createDealerHandler = async (req: Request, res: Response) => {
     console.log(req.body);
-    try { /**MongoDb call */
-        const dealer = await DealerModel.create(req.body)
-        return res.send(dealer)
-    } catch (e: any) {
-        logger.error(e);
-        return res.status(409).send(e.message);
+    if (req.body.auth) {
+        try { /**MongoDb call */
+            const dealer = await DealerModel.create(req.body)
+            return res.send(dealer)
+        } catch (e: any) {
+            logger.error(e);
+            return res.status(409).send(e.message);
+        }
+    } else {
+        return res.status(401).send("not authorised");
     }
 };
 
@@ -25,39 +29,47 @@ const getAllDealersHandler = async (req: Request, res: Response) => {
 };
 
 const updateDealerHandler = async (req: Request, res: Response) => {
-    const id = req.body._id;
-    const data = req.body
-    try {
-        /**MongoDb call */
-        let updatedDealer
-        if (mongoose.Types.ObjectId.isValid(id)) {
-            updatedDealer = await DealerModel.findByIdAndUpdate(id, { $set: { name: data.name, email: data.email, company: data.company } }, { new: true })
-            if (updatedDealer) {
-                return res.send(updatedDealer)
+    if (req.body.auth) {
+        const id = req.body._id;
+        const data = req.body
+        try {
+            /**MongoDb call */
+            let updatedDealer
+            if (mongoose.Types.ObjectId.isValid(id)) {
+                updatedDealer = await DealerModel.findByIdAndUpdate(id, { $set: { name: data.name, email: data.email, company: data.company } }, { new: true })
+                if (updatedDealer) {
+                    return res.send(updatedDealer)
+                }
             }
-        }
 
-    } catch (e: any) {
-        logger.error(e);
-        return res.status(409).send(e.message);
+        } catch (e: any) {
+            logger.error(e);
+            return res.status(409).send(e.message);
+        }
+    } else {
+        return res.status(401).send("not authorised");
     }
 };
 const deleteDealerHandler = async (req: Request, res: Response) => {
-    const id = req.body._id;
-    try {
-        /**MongoDb call */
-        let deletedDealer
-        if (mongoose.Types.ObjectId.isValid(id)) {
-            deletedDealer = await DealerModel.findByIdAndRemove(id)
-            if (deletedDealer) {
-                return res.send(deletedDealer)
-            } else {
-                return res.send("no such dealer exits")
+    if (req.body.auth) {
+        const id = req.body._id;
+        try {
+            /**MongoDb call */
+            let deletedDealer
+            if (mongoose.Types.ObjectId.isValid(id)) {
+                deletedDealer = await DealerModel.findByIdAndRemove(id)
+                if (deletedDealer) {
+                    return res.send(deletedDealer)
+                } else {
+                    return res.send("no such dealer exits")
+                }
             }
+        } catch (e: any) {
+            logger.error(e);
+            return res.status(409).send(e.message);
         }
-    } catch (e: any) {
-        logger.error(e);
-        return res.status(409).send(e.message);
+    } else {
+        return res.status(401).send("not authorised");
     }
 };
